Add some bash scripts to setup tmux sessions.
I reworked the rpp_web_app.sh script to be cleaner and to handle
checking for existance of the session before creating it and attaching
if it already exists. I used that pattern for the other tmux session
startup scripts as well. I added a script for my personal blog and for
octopusci.