Set the left status width and enable utf8.
I set the left status width because most of my session names are pretty
long and I wanted to actually be able to see them. It might be a little
long at the moment but I can trim it down over time as I see fit. The
utf8 status option allows the use of utf8 characters in the status bar
at the bottom.