Removed the rpp_web_app.sh tmux setup script.
I did this because I have now moved this script into the rpp_web_app project
git repository so that it can be shared with the other developers of the
rpp_web_app project. I also modified it to be written in a generic way rather
than with specifics to my directory structure, etc.