Added option to bring ssh-agent envs over.
I did this so that when windows are created in tmux the ssh-agent environment
variable values are brought over from the initial launching terminal. This
allows ssh calls or ssh calls via git to use the ssh-agent that is running
normally on Mac OS X which ties into the keychain and all of that.