Cleanup the utility scripts
This involved changing the sourcing of the shared tmux-utils.sh file
which I thought was source relative to the file but it was sourced
relative to execution. So I changed it to be relative to the home
directory which is the installation path so that it works no matter
where the user executes the utility from.

I also added a section to README.md so that I can remember what I need
to do when setting it up again.

Additionally I removed the appfit.sh script as it was garbarage that was
never really implemented. Having noise around just makes things
confusing so might as well clean up the cruft.