Add more kinly sessions to kinly-tmux
I made this change to add 3 new sessions to the kinly-tmux setup script
so that I would have a session for cbkc-csv-anonymize,
mx-transaction-webhook-lambdas, and trans-faktory-server.

I did this because I want the kinly-tmux script to bring up all the
sessions for all the kinly repositories.

I am doing this by simply calling the create_session function 3 more
times, one for each of the respective sessions with the pathing and
naming info.