Make git-ps-tmux setup single session
Make git-ps-tmux setup single session so that I don't have so much noise
in the sessions list and instead have a single session with respective
windows for the various parts of the git-patch-stack project overall.
This also gives me an easy way to kill the entire git-patch-stack
session if I want to get rid of it. Beyond that it also sets up the
website and book commands to render them respectively and watch the
files for changes.

This will help with me simply forgetting the commands as it will start
with them in there and then I can look through the command history to
find them if I forget within a session.